Fix signal 11 during compilation of profile data.  Was caused by a memory corruption error that was produced in computeWorkHistogram
git-svn-id: svn+ssh://lc-svn.llnl.gov/usr/gapps/emhm/svn/Cardioid/EP/trunk@1118 baecf98e-1e81-4516-8920-d4c889e67ea4

+++ src/assignCellsToTasks.cc
@@ -662,7 +662,7 @@ namespace
 
          for (int ip=0; ip<nProcs; ip++)
          {
-            int pwork = pework_all[ip];
+            double pwork = pework_all[ip];
             int bin = (pwork-minwork)/delta;
             phist[bin]++;
          }
